I added new features to the work.py to be able to easily create exe

+++ work.py
@@ -20,8 +20,9 @@ def generateExecutable(*args):
 		'index.py', "--onefile"
 	])
 	shutil.copytree("./dependencies/", "./dist/dependencies/")
-	shutil.rmtree("./build")
-	os.remove("index.spec")
+	if ("-keep" not in args):
+		shutil.rmtree("./build")
+		os.remove("index.spec")
 
 def createview(name="Page", *args):
 	data = snips.view(name)
